requirements.txt
streamlit==1.28.1
torch==2.7.1
torchvision==0.22.2
opencv-python==4.12.0.88
Pillow==10.3.0
numpy==2.2.6
pandas==2.3.1
click==8.1.7
pytest==8.3.4
pytest-cov==4.1.0
